pyproject.toml
[project]
authors = [
  {name = "Patrick Arminio", email = "patrick.arminio@gmail.com"},
]
dependencies = [
    "django>=4.1",
    "strawberry-graphql==0.125.0.dev1660322955",
    "channels>=3.0.5",
]
description = ""
license = {text = "MIT"}
name = ""
requires-python = ">=3.10"
version = ""
[project.optional-dependencies]

[build-system]
build-backend = "pdm.pep517.api"
requires = ["pdm-pep517>=1.0.0"]

[tool]
[tool.pdm]

[tool.pdm.scripts]
server = "python manage.py runserver"
[tool.pdm.dev-dependencies]
dev = [
    "black>=22.6.0",
    "flake8>=5.0.4",
    "mypy>=0.971",
    "pdbpp>=0.10.3",
]

[tool.mypy]
plugins = ["strawberry.ext.mypy_plugin"]
